Generate a unique semaphore key by including size of semaphore set (#113)
If we increase the size of the semaphore set, we will see errors.

This is because calls to read from the set require you to specify the
cardinality of the semaphore set. If this changes, the ID is invalid.

The solution here is to include the set cardinality in the key before
we SHA hash it.