Fix the 'acquire_with_fork' test
Under the existing implementation, a forked process attempts to make an assertion.

The results of this minitest assertion will always be ignored by the parent process,
because the master process doesn't care about the exit status of the forked process.

To work around this, the exit status of the child process is used to determine if a
timeout has occured, and the assertion is run in the parent process, running the rest
of the test suite.