Add support for a minimum number of tickets when using the quota strategy
Since the quota strategy makes the number of tickets available proportional to the number of workers,
it's possible that the number of tickets may be very constrained during some normal operations,
such as rolling restarts during, for instance, a deploy.

To prevent these situations from resulting in a large number of tripped circuit breakers,
a minimum number of tickets can be specified.

The actual ticket count used will be the max of the tickets calculated by the quota, and the
minimum value.