Register all workers - not just quota workers
Currently it's possible to have spurious timeouts when transitioning
between static and quota-based ticket allocation.

This is primarily a problem when switching from static to quota-based,
as there will be a large number of unregistered static-ticket based
workers, who will have a very small number of tickets as the
quota-ticket based workers boot and register.

By registering all workers, we can account for this issue.