Add ability to unregister all resources, and clear all semian_resources (#134)
* Add ability to unregister all resources, and clear all semian_resources

Functionality to track references to resource consumers - instances of
semian adaptors - has been added. Any time an instance of a semian adaptor
retrieves a semian resource, it will register itself as a consumer of the resource.

This allows for the resource reference to be completely cleared when the resource
is unregistered. To avoid this preventing GC on the semian adaptor object, this
reference is intentionally weak.

* Don't unregister if there is nothing registered