Make sure internal semaphore release is protected.
All of the resource semaphores are used (this is common in our job servers). We
still use a number of tickets proportional to the number of workers with rescue.
This causes a lot of thrashing on the semaphore limits on every deploy. In
particular, in this case, a worker wanted to reduce the number of tickets
because it was getting shut down.

* It grabbed the internal semaphore.
* Timed out as it tried to reduce the count.
* Raised.

Since the internal semaphore was not protected, it was never released, so the
count could not be further adjusted.