Handle new Rails connection changes
The refactoring in https://github.com/rails/rails/pull/44591 we started
seeing errors like `NoMethodError: undefined method 'semian_resource'
for nil:NilClass` coming from Semian. While this is called from
shopify/shopify, the `connection` that's `nil` is the one in the
`rails.rb` monkey patch.

After debugging with matthewd we found that Semian is relying on the
previous connection being available, whereas the new AR behavior means
it might not be.

To fix this we can memoize the `semian_resource` on `connection` in the
adapter, and then when the connection is reset or disconencted, set
`semian_resource` to `nil` and grab the new connection instance next
time we need it.

In addition to these changes I decided to add `activerecord` as a
development dependency to the gemspec so that we are able to test these
changes. Previously you would need to bundle open the gem in
shopify/shopify to test changes before sending a PR. It's important that
we're able to test what Semian is relying on so it makes sense to have a
depenedency on Active Record here.

To test the Rails upgrade locally I cloned Rails into my workspace and
added `gem "activerecord", path: "/workspace/rails"` to the Gemfile.
This let me easily debug the changes once I understood what was breaking
in shopify/shopify. Hopefully this makes debugging connections easier in
the future.