WIP: Adapter for GitHub's Trilogy MySQL client
This puts the basis of an adapter, it's still missing
quite a few things.

We might also need to add a few things upstream, such as
`Trilogy#closed?` etc.

Also Trilogy raise a lot of different base exceptions, so
it's very tricky to list them all. We might need to resort to
consider all `IOError` and `SystemCallError` subclasses.

Co-Authored-By: adrianna.chang@shopify.com