Use ActiveRecord::ConnectionNotEstablished as base error for TrilogyAdapter
One part of the design of semian is that the exception it raises should
ideally be caught by naive `rescue SomeClient::BaseConnectionError`.

The reason is that semian raise a fast error when it thinks it's likely
using the resource right now would lead to a slow network error of
some kind.