Support only Ruby 3.2 and onward
We noticed that Dependabot has stopped being able to issue updates due to the gemspec specifying Ruby 2.7+, but we have
optional dependencies that require ruby 3.1+. While our `.ruby-version` file specifies 3.2, Dependabot only reads the
gemspec and not the ruby-version file. This is a known bug (https://github.com/dependabot/dependabot-core/issues/10108)
that has not been addressed.

Looking at Ruby's [deprecation schedule](https://www.ruby-lang.org/en/downloads/branches/), 2.7 was EOL in 2023 and 3.1
will be EOL in a month. This gives us good reason to support 3.2+ going forward, bringing the gemspec in line with our
testing versions.

In this, we also needed to bump dependencies so they would work with newer ruby versions. In this, we ran into gRPC
tests not exiting despite passing, so a lot of defensiveness has been added to those tests to ensure that we're actually
closing everything and cleaning up after ourselves. In addition, we noticed that when running tests locally with
docker-compose, gRPC would return `GRPC::Unavailable` instead of `GRPC::DeadlineExceeded`

- Bump gemspec to use Ruby 3.2+
- Drop 2.7 and 3.0 from testing, add 3.3 and 3.4
- Bump toxiproxy test container from 2.5.0 -> 2.12.0
- Bump gRPC to 1.71.0
- Bump trilogy to 2.9
- Update lockfiles
- Add timeouts in CI
- Update rubocop to use plugins
- Style fixes
- Fix gRPC test to exit cleanly.