Attempt acquiring semaphore before releasing GVL
Particularly when there are multiple threads waiting to run, there is a cost to releasing the GVL that outweighs the
benefit for short system calls.

This commit changes acquire_semaphore_without_gvl to attempt to acquire the semaphore while also including IPC_NOWAIT in
flags, which should return immediately if the request isn't available.